feat: add overlay toggling by cmd + shift + p
